feat(api): Adds unstable_pages module to JS API
