+++ .changeset/green-planets-wink.md
@@ -1,0 +1,5 @@
+---
+"wrangler": patch
+---
+
+Adds unstable_pages module to JS API
