+++ packages/wrangler/src/api/index.ts
@@ -1,2 +1,3 @@
 export { unstable_dev } from "./dev";
 export type { UnstableDevWorker, UnstableDevOptions } from "./dev";
+export { unstable_pages } from "./pages";
