+++ packages/wrangler/src/api/pages/index.ts
@@ -1,0 +1,5 @@
+import { publish } from "./publish";
+
+export const unstable_pages = {
+	publish,
+};
